move NoopPlan to common package so StressJobServiceBench can import
