+++ job/common/src/main/java/alluxio/job/plan/NoopPlanConfig.java
@@ -9,9 +9,7 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.job;
-
-import alluxio.job.plan.PlanConfig;
+package alluxio.job.plan;
 
 import com.google.common.base.MoreObjects;
