+++ job/server/src/main/java/alluxio/job/plan/NoopPlanDefinition.java
@@ -12,7 +12,6 @@
 package alluxio.job.plan;
 
 import alluxio.collections.Pair;
-import alluxio.job.NoopPlanConfig;
 import alluxio.job.RunTaskContext;
 import alluxio.job.SelectExecutorsContext;
 import alluxio.job.util.SerializableVoid;
